incorporate gdd

+++ historical_weather/Scripts/weatherHistoryCronMaster.R
@@ -38,7 +38,18 @@ if(lockCheck ==0 & freeSpace>10000){ #not another one of these tasks running and
 		  cat(paste("finished, removing lockfile: ",Sys.time(),"\n",sep=""),file=paste(homeDir,"/log.txt",sep=""),append=T)
 		  
 		  setwd(homeDir)
-		  unlink("lockfileFolder",recursive=T)	
+		  
+		  calculateAndEmailGDDProgress = tryCatch(source("../gddGetter/Scripts/gdd_getter.R"),error=function(e) e)
+		  setwd(homeDir)
+
+		  if(inherits(calculateAndEmailGDDProgress,"error")==F){
+		  	 cat(paste("error with gdd progress update: ",Sys.time(),"\n",sep=""),file=paste(homeDir,"/log.txt",sep=""),append=T)
+
+		  }else{
+ 			cat(paste("finished, removing lockfile: ",Sys.time(),"\n",sep=""),file=paste(homeDir,"/log.txt",sep=""),append=T)
+
+		  	unlink("lockfileFolder",recursive=T)	
+		  }
 		}else{
 		  setwd(homeDir)
 		  errorText = paste("cron task call script had an error on uploading, blocking future tasks until resolved: ",Sys.time(),"\n",sep="")
